+++ README.md
@@ -35,10 +35,10 @@ We are planning to address this problem in the future.
 
 ## Usage
 
-First, install opencensusd if you haven't.
+First, install ocagent if you haven't.
 
 ```
-$ go get github.com/census-instrumentation/opencensus-service/cmd/opencensusd
+$ go get github.com/census-instrumentation/opencensus-service/cmd/ocagent
 ```
 
 Create a config.yaml file in the current directory and modify
@@ -64,14 +64,14 @@ to the daemon if it is running.
 $ go run "$(go env GOPATH)/src/github.com/census-instrumentation/opencensus-service/example/main.go"
 ```
 
-Run opencensusd:
+Run ocagent:
 
 ```
-$ opencensusd
+$ ocagent
 ```
 
 You should be able to see the traces in Stackdriver and Zipkin.
-If you stop the opencensusd, example application will stop exporting.
+If you stop the ocagent, example application will stop exporting.
 If you run it again, it will start exporting again.
 
 ## OpenCensus Agent
